--- core-design-system/src/main/res/drawable/ic_tab_home.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:drawable="@drawable/ic_tab_home_n" android:state_checked="false"/>
-    <item android:drawable="@drawable/ic_tab_home_f" android:state_checked="true"/>
-</selector>
