--- core-design-system/src/main/res/drawable/ic_tab_home_d.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="20dp"
-    android:height="21dp"
-    android:viewportWidth="20"
-    android:viewportHeight="21">
-  <path
-      android:pathData="M19.069,7.204L10.622,0.487C10.445,0.346 10.226,0.27 10,0.27C9.774,0.27 9.555,0.346 9.378,0.487L0.93,7.204C0.694,7.391 0.504,7.629 0.373,7.9C0.243,8.171 0.175,8.468 0.175,8.769V18.187C0.175,18.824 0.428,19.434 0.878,19.884C1.328,20.334 1.938,20.587 2.575,20.587H8V15.835C8,15.57 8.105,15.316 8.293,15.128C8.48,14.94 8.735,14.835 9,14.835H11C11.265,14.835 11.519,14.94 11.707,15.128C11.894,15.316 12,15.57 12,15.835V20.587H17.424C18.06,20.587 18.671,20.334 19.121,19.884C19.571,19.434 19.824,18.824 19.824,18.187V8.77C19.824,8.469 19.756,8.172 19.625,7.901C19.494,7.63 19.304,7.392 19.069,7.205"
-      android:fillColor="#B0B8C1"
-      android:fillType="evenOdd"/>
-</vector>
