--- core-design-system/src/main/res/drawable/ic_tab_my_d.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="25dp"
-    android:viewportWidth="24"
-    android:viewportHeight="25">
-  <path
-      android:pathData="M16.887,6.911C16.887,7.579 16.761,8.241 16.515,8.859C16.269,9.476 15.908,10.038 15.454,10.51C14.999,10.983 14.46,11.358 13.866,11.614C13.272,11.87 12.635,12.002 11.993,12.002C10.694,12.002 9.449,11.466 8.531,10.511C7.613,9.557 7.097,8.262 7.097,6.912C7.097,6.243 7.224,5.581 7.47,4.963C7.716,4.346 8.076,3.785 8.531,3.312C9.448,2.357 10.693,1.82 11.992,1.82C13.29,1.82 14.535,2.356 15.453,3.311C16.371,4.266 16.887,5.56 16.887,6.911ZM11.992,13.559C4.943,13.559 2.201,18.225 2.201,20.395C2.201,22.565 8.038,23.143 11.992,23.143C15.946,23.143 21.783,22.565 21.783,20.395C21.783,18.225 19.041,13.559 11.992,13.559Z"
-      android:fillColor="#B0B8C1"
-      android:fillType="evenOdd"/>
-</vector>
