+++ presentation/history/.gitignore
