+++ presentation/home/src/main/java/com/depromeet/threedays/home/MainActivity.kt
@@ -5,8 +5,6 @@ import androidx.fragment.app.Fragment
 import com.depromeet.threedays.core.BaseActivity
 import com.depromeet.threedays.home.databinding.ActivityMainBinding
 import com.depromeet.threedays.home.home.HomeFragment
-import com.depromeet.threedays.my.MyFragment
-import com.depromeet.threedays.statistics.StatisticsFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -24,12 +22,6 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main) {
                 when (it.itemId) {
                     R.id.homeFragment -> {
                         changeFragment(HomeFragment())
-                    }
-                    R.id.statisticsFragment -> {
-                        changeFragment(StatisticsFragment())
-                    }
-                    R.id.myFragment -> {
-                        changeFragment(MyFragment())
                     }
                 }
                 true
