+++ presentation/home/src/main/res/drawable/ic_tab_mate.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:drawable="@drawable/ic_tab_mate_unselected" android:state_checked="false"/>
+    <item android:drawable="@drawable/ic_tab_mate_selected" android:state_checked="true"/>
+</selector>
