--- presentation/my/src/main/java/com/depromeet/threedays/my/MyViewModel.kt
@@ -1,6 +1,0 @@
-package com.depromeet.threedays.my
-
-import com.depromeet.threedays.core.BaseViewModel
-
-class MyViewModel : BaseViewModel() {
-}
