--- presentation/statistics/build.gradle
@@ -1,52 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'org.jetbrains.kotlin.android'
-    id "kotlin-kapt"
-    id 'dagger.hilt.android.plugin'
-}
-
-android {
-    compileSdk 32
-
-    defaultConfig {
-        minSdk 26
-        targetSdk 32
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
-    }
-    kotlinOptions {
-        jvmTarget = '1.8'
-    }
-    buildFeatures {
-        dataBinding true
-    }
-}
-
-dependencies {
-
-    implementation(project(":core"))
-    implementation(project(":navigator"))
-    implementation(project(":domain"))
-    implementation(project(":core-design-system"))
-
-    implementation(jetpackDeps)
-    implementation(coroutines)
-
-    implementation deps.hilt.core
-    kapt deps.hilt.compiler
-
-    testImplementation(testDeps)
-    androidTestImplementation(androidTestDeps)
-}
