--- presentation/statistics/src/main/java/com/depromeet/threedays/statistics/StatisticsFragment.kt
@@ -1,12 +1,0 @@
-package com.depromeet.threedays.statistics
-
-import androidx.fragment.app.viewModels
-import com.depromeet.threedays.core.BaseFragment
-import com.depromeet.threedays.statistics.databinding.FragmentStatisticsBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class StatisticsFragment: BaseFragment<FragmentStatisticsBinding, StatisticsViewModel>(R.layout.fragment_statistics) {
-    override val viewModel by viewModels<StatisticsViewModel>()
-
-}
