--- presentation/statistics/src/main/java/com/depromeet/threedays/statistics/StatisticsViewModel.kt
@@ -1,6 +1,0 @@
-package com.depromeet.threedays.statistics
-
-import com.depromeet.threedays.core.BaseViewModel
-
-class StatisticsViewModel : BaseViewModel() {
-}
