+++ settings.gradle
@@ -22,5 +22,6 @@ include ':domain'
 include ':presentation'
 include ':presentation:home'
 include ':presentation:register'
-include ':presentation:my'
-include ':presentation:statistics'
+include ':presentation:history'
+include ':presentation:mypage'
+include ':presentation:mate'
